Modularize kickstart nvim config for better maintenance in the future
